requirements.txt
Django
django-crispy-forms
django-allauth
djangorestframework
django-channels
celery
redis
PyGithub
django-tailwind
django-imagekit
mysqlclient